day2/requirements.txt
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
beautifulsoup4
lxml
requests
pypdf
